Upgraded for new PSMoveAPI and modern Pi OS
Uses new PSMoveAPI v4
Uses SystemD to run on boot
New API doesn't require original bluetooth hacks
Remembers controllers so pairing not needed every time
Orbs now glow orange when connected to the game lobby so they don't look broken
