--- supervisor/conf.d/joust.conf
@@ -1,13 +1,0 @@
-[program:scan_enable]
-command=/home/pi/scan_enable.sh
-user=root
-autorestart=true
-startretries=1000
-
-[program:oust]
-command=/usr/bin/python /home/pi/psmoveapi/build/oust.py
-directory=/home/pi/psmoveapi/build
-environment=HOME="/home/pi"
-user=root
-autorestart=true
-startretries=1000
